refactor(racing): RandomNumberGenerator 랜덤 숫자 타입 및 생성 로직 변경
- 랜덤 숫자 타입 수정
  - double -> int
- 생성 로직 변경
  - Math.random() 활용 -> Random.nextInt() 활용
- generate()한 값을 사용하는 로직의 parameter type 수정

+++ src/main/java/racing/domain/Car.java
@@ -1,7 +1,7 @@
 package racing.domain;
 
 public class Car implements Comparable {
-    private final double MOVE_LOWER_BOUND = 4.0;
+    private final int MOVE_LOWER_BOUND = 4;
     private final String name;
     private final String PROGRESS_INNER_STRING = " : ";
     private final String PROGRESS_SYMBOL = "-";
@@ -15,7 +15,7 @@ public class Car implements Comparable {
         return name;
     }
 
-    public boolean move(double number) {
+    public boolean move(int number) {
         boolean isMoving = isMoving(number);
         if (isMoving) {
             increaseDistance();
@@ -27,7 +27,7 @@ public class Car implements Comparable {
         distance++;
     }
 
-    private boolean isMoving(double number) {
+    private boolean isMoving(int number) {
         return number >= MOVE_LOWER_BOUND;
     }
